Synapses AI Challenge/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
aiohttp==3.9.1
google-search-results==2.4.2
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.4
numpy==1.25.2
PyPDF2==3.0.1 